♻️🎪 Split application page
Split the route file to improve readability.
